refactor: organize into modules
