+++ client/src/app/app.module.ts
@@ -1,48 +1,26 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
-import { ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
-import { MatButtonModule } from '@angular/material/button';
-import { MatNativeDateModule } from '@angular/material/core';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatPaginatorModule } from '@angular/material/paginator'
-import { MatSortModule } from '@angular/material/sort'
-import { MatTableModule } from '@angular/material/table'
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { PolicyComponent } from './policy/policy.component';
 import { ApiService } from './services/api.service';
-import { PolicyListComponent } from './policy-list/policy-list.component';
-import { ClickableDirective } from './components/clickable.directive';
-import { CanDeactivateRouteService } from './can-deactivate-route.service';
+import { CanDeactivateRouteService } from './services/can-deactivate-route.service';
+import { PolicyModule } from './modules/policy/module';
+import { SharedModule } from './components/module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    ClickableDirective,
-    PolicyComponent,
-    PolicyListComponent
   ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
     HttpClientModule,
-    ReactiveFormsModule,
-    MatButtonModule,
-    MatDatepickerModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatNativeDateModule,
-    MatSnackBarModule,
-    MatPaginatorModule,
-    MatSortModule,
-    MatTableModule,
-    AppRoutingModule
+    AppRoutingModule,
+    PolicyModule,
+    SharedModule,
   ],
   providers: [
     ApiService,
