+++ client/src/app/components/module.ts
@@ -1,0 +1,49 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+import { MatButtonModule } from '@angular/material/button';
+import { MatNativeDateModule } from '@angular/material/core';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatPaginatorModule } from '@angular/material/paginator'
+import { MatSortModule } from '@angular/material/sort'
+import { MatTableModule } from '@angular/material/table'
+
+import { ClickableDirective } from './clickable.directive';
+
+@NgModule({
+    declarations: [
+        ClickableDirective,
+    ],
+    imports: [
+        BrowserModule,
+        ReactiveFormsModule,
+        MatButtonModule,
+        MatDatepickerModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatNativeDateModule,
+        MatSnackBarModule,
+        MatPaginatorModule,
+        MatSortModule,
+        MatTableModule,
+    ],
+    exports: [
+        BrowserModule,
+        ReactiveFormsModule,
+        MatButtonModule,
+        MatDatepickerModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatNativeDateModule,
+        MatSnackBarModule,
+        MatPaginatorModule,
+        MatSortModule,
+        MatTableModule,
+
+        ClickableDirective,
+    ],
+})
+export class SharedModule { }
