+++ client/src/app/modules/policy/components/item/policy.component.scss
