+++ client/src/app/modules/policy/components/list/policy-list.component.scss
