+++ client/src/app/modules/policy/module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+
+import { PolicyComponent } from './components/item/policy.component';
+import { PolicyListComponent } from './components/list/policy-list.component';
+import { RouterModule } from '@angular/router';
+import { routes } from './routes';
+import { SharedModule } from 'src/app/components/module';
+
+@NgModule({
+    declarations: [
+        PolicyComponent,
+        PolicyListComponent,
+    ],
+    imports: [
+        SharedModule,
+        RouterModule.forChild(routes),
+    ],
+})
+export class PolicyModule { }
