+++ client/src/app/modules/policy/services/policy.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { PolicyItem, PolicyListItem } from 'src/shared/types';
-import { ApiService } from './api.service';
+import { ApiService } from 'src/app/services/api.service';
 
 @Injectable({
   providedIn: 'root'
