+++ client/src/app/services/can-deactivate-route.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { CanDeactivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
 
-import { ActiveComponent } from './components/active-component.service';
+import { ActiveComponent } from '../components/active-component.service';
 
 @Injectable({
   providedIn: 'root'
